refactor: move Gradle workspace discovery to provider layer
Extract ~120 lines of Gradle-specific code (init script, subproject
parsing, binary resolution) from ExhortApi into a dedicated
GradleWorkspaceDiscovery utility class in the provider layer,
following the same pattern as JsWorkspaceDiscovery.

Add wrapper-preference tests for gradlew resolution path.
